+++ vscode/src/context/openctx/remoteRepositorySearch.ts
@@ -1,6 +1,6 @@
-import { graphqlClient, isError } from '@sourcegraph/cody-shared'
+import { contextFiltersProvider, graphqlClient, isError } from '@sourcegraph/cody-shared'
 
-import type { Item, Mention, Provider } from '@openctx/client'
+import type { Item, Provider } from '@openctx/client'
 
 const RemoteRepositorySearch: Provider & {
     providerUri: string
@@ -8,16 +8,12 @@ const RemoteRepositorySearch: Provider & {
     providerUri: 'internal-remote-repository-search',
 
     meta() {
-        return { name: 'Sourcegraph Repositories', mentions: {} }
+        return { name: 'Remote Repositories', mentions: {} }
     },
 
     async mentions({ query }) {
-        if (query && query.length < 3) {
-            return []
-        }
-
         try {
-            const dataOrError = await graphqlClient.searchRepos(10, undefined, query)
+            const dataOrError = await graphqlClient.searchRepos(30, undefined, query)
 
             if (isError(dataOrError) || dataOrError === null) {
                 return []
@@ -25,16 +21,18 @@ const RemoteRepositorySearch: Provider & {
 
             const repositories = dataOrError.repositories.nodes
 
-            return repositories.map(
-                repo =>
-                    ({
-                        uri: repo.url,
-                        title: repo.name,
-                        data: {
-                            repoId: repo.id,
-                        },
-                    }) as Mention
-            )
+            return repositories.map(repo => ({
+                uri: repo.url,
+                title: repo.name,
+                // By default we show <title> <uri> in the mentions menu.
+                // As repo.url and repo.name are almost same, we do not want to show the uri.
+                // So that is why we are setting the description to " " string.
+                description: ' ',
+                data: {
+                    repoId: repo.id,
+                    isIgnored: contextFiltersProvider.isRepoNameIgnored(repo.name),
+                },
+            }))
         } catch (error) {
             return []
         }