+++ vscode/src/non-stop/FixupTask.ts
@@ -10,7 +10,6 @@ import {
 
 import type { EditIntent, EditMode } from '../edit/types'
 
-import { getOverridenModelForIntent } from '../edit/utils/edit-models'
 import type { FixupFile } from './FixupFile'
 import type { Diff } from './diff'
 import { CodyTaskState } from './utils'
@@ -84,7 +83,6 @@ export class FixupTask {
         this.id = Date.now().toString(36).replaceAll(/\d+/g, '')
         this.instruction = instruction.replace(/^\/(edit|fix)/, ps``).trim()
         this.originalRange = selectionRange
-        this.model = getOverridenModelForIntent(this.intent, this.model)
     }
 
     /**