+++ vscode/src/tree-sitter/queries.ts
@@ -1,7 +1,13 @@
 import type { SupportedLanguage } from './grammars'
+import { cQueries } from './queries/c'
+import { cppQueries } from './queries/cpp'
 import { goQueries } from './queries/go'
+import { javaQueries } from './queries/java'
 import { javascriptQueries } from './queries/javascript'
+import { kotlinQueries } from './queries/kotlin'
+import { phpQueries } from './queries/php'
 import { pythonQueries } from './queries/python'
+import { rustQueries } from './queries/rust'
 
 export type QueryName =
     | 'singlelineTriggers'
@@ -43,4 +49,10 @@ export const languages: Partial<Record<SupportedLanguage, Record<QueryName, stri
     ...javascriptQueries,
     ...goQueries,
     ...pythonQueries,
+    ...javaQueries,
+    ...kotlinQueries,
+    ...phpQueries,
+    ...rustQueries,
+    ...cQueries,
+    ...cppQueries,
 } as const